Implement simple plugin system on top of entry points
The module provides simple base class for plugins and
a plugin manager for the dynamic inventory.

All plugins are required to subclass the base class provided.
The plugin manages plugin activation based on an entry in
the configuration file named after the plugin.

The plugin manager or any other part of the code
does not check for conflicts in the results of multiple
plugins.