Implement updates testing and copr plugins
The two plugins are used to generate metadata for hosts
in the inventory. The plugins use the entry point machinery
to get loaded even though they are a part of the package.

The updates testing plugin provides a switch for using the updates testing
repository in the installation.

The COPR plugin inserts an array of copr repositories to enable on
the hosts.

The plugins need entries in the configuration file to be activated.