Implement the Inventory class
The Inventory class builds a representation of an ansible
inventory based on hosts taken from the configuration file
of FreeIPA upstream integation tests.

The hosts are sorted into several groups based on their role and
provided with metadata about required packages for each role.

The inventory also invokes any plugins registered to the package
entry point and applies the data to the hostvars for each particular
host.